+++ logprep/connector/confluent_kafka/input.py
@@ -37,7 +37,7 @@ from typing import Optional, Tuple, Union
 
 import msgspec
 from attrs import define, field, validators
-from confluent_kafka import (
+from confluent_kafka import (  # type: ignore
     OFFSET_BEGINNING,
     OFFSET_END,
     OFFSET_INVALID,
@@ -47,7 +47,7 @@ from confluent_kafka import (
     Message,
     TopicPartition,
 )
-from confluent_kafka.admin import AdminClient
+from confluent_kafka.admin import AdminClient  # type: ignore
 
 from logprep.abc.connector import Connector
 from logprep.abc.input import (
@@ -331,6 +331,7 @@ class ConfluentKafkaInput(Input):
         """
 
         stats = self._decoder.decode(stats)
+        assert isinstance(stats, dict)
         self.metrics.librdkafka_age += stats.get("age", DEFAULT_RETURN)
         self.metrics.librdkafka_rx += stats.get("rx", DEFAULT_RETURN)
         self.metrics.librdkafka_tx += stats.get("tx", DEFAULT_RETURN)
@@ -395,7 +396,7 @@ class ConfluentKafkaInput(Input):
         base_description = super().describe()
         return f"{base_description} - Kafka Input: {self._config.kafka_config['bootstrap.servers']}"
 
-    def _get_raw_event(self, timeout: float) -> bytearray:
+    def _get_raw_event(self, timeout: float) -> bytes | None:
         """Get next raw Message from Kafka.
 
         Parameters
@@ -429,7 +430,7 @@ class ConfluentKafkaInput(Input):
         self.metrics.current_offsets.add_with_labels(message.offset() + 1, labels)
         return message.value()
 
-    def _get_event(self, timeout: float) -> Union[Tuple[None, None], Tuple[dict, dict]]:
+    def _get_event(self, timeout: float) -> Tuple[None, None] | Tuple[dict, bytes]:
         """Parse the raw document from Kafka into a json.
 
         Parameters