+++ logprep/connector/http/output.py
@@ -13,7 +13,7 @@ An example config file would look like:
     :linenos:
 
     output:
-      myhttpoutput:
+      my_http_output:
         type: http_output
         target_url: http://the.target.url:8080
         username: user
@@ -120,6 +120,12 @@ class HttpOutput(Output):
         """URL of the endpoint that receives the events"""
         timeout: int = field(validator=validators.instance_of(int), default=2)
         """Timeout in seconds for the http request"""
+        verify: bool | str = field(
+            validator=validators.instance_of(str | bool),
+            default=True,
+            converter=lambda x: "" if x is None else x,
+        )
+        """Switch to disable ssl verification or path to certificate"""
 
     @property
     def user(self):
@@ -139,6 +145,11 @@ class HttpOutput(Output):
     @cached_property
     def _headers(self):
         return {"Content-Type": "application/x-ndjson; charset=utf-8"}
+
+    @property
+    def verify(self):
+        """Return the ssl verify status that is used for the http request"""
+        return self._config.verify
 
     @property
     def statistics(self) -> str:
@@ -163,13 +174,14 @@ class HttpOutput(Output):
 
     def store(self, document: tuple[str, dict | list[dict]] | dict) -> None:
         if isinstance(document, tuple):
-            target, document = document
-            target = f"{self._config.target_url}{target}"
+            target, payload = document
+            target = f"{self._config.target_url}/{target}"
+            self.store_custom(payload, target)
         else:
             target = self._config.target_url
-        self.store_custom(document, target)
+            self.store_custom(document, target)
 
-    def store_custom(self, document: dict | tuple | list, target: str) -> None:
+    def store_custom(self, document: dict | tuple | list | str, target: str) -> None:
         """Send a post request with given data to the specified endpoint"""
         if isinstance(document, (tuple, list)):
             request_data = self._encoder.encode_lines(document)
@@ -177,6 +189,9 @@ class HttpOutput(Output):
         elif isinstance(document, dict):
             request_data = self._encoder.encode(document)
             document_count = 1
+        elif isinstance(document, str):
+            document_count = document.count(";") + 1
+            request_data = document.replace(";", "\n")
         else:
             error = TypeError(f"Document type {type(document)} is not supported")
             self.metrics.number_of_failed_events += 1
@@ -184,16 +199,15 @@ class HttpOutput(Output):
             return
         try:
             try:
-                logger.debug(request_data)
                 response = requests.post(
                     url=target,
                     headers=self._headers,
-                    verify=False,
+                    verify=self.verify,
                     auth=(self.user, self.password),
                     timeout=(self.timeout, self.timeout),
                     data=request_data,
                 )
-                logger.debug("Servers response code is: %i", response.status_code)
+                # logger.debug("Servers response code is: %i", response.status_code)
                 self.metrics.status_codes.add_with_labels(
                     1,
                     {